requirements.txt
# ============================================================================
# 회전교차로 상호작용 예측 연구 프로젝트 - 의존성 목록
# ============================================================================

# ----------------------------------------------------------------------------
# Core Deep Learning Frameworks
# ----------------------------------------------------------------------------
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# ----------------------------------------------------------------------------
# Graph Neural Networks
# ----------------------------------------------------------------------------
torch-geometric>=2.3.0
# torch-geometric-temporal>=0.54.0  # 선택사항 (A3TGCN 사용 시 필요, 빌드 필요)

# PyTorch Geometric 의존성 (명시적 설치)
xxhash>=3.0.0
aiohttp>=3.8.0
psutil>=5.8.0
requests>=2.28.0

# ----------------------------------------------------------------------------
# Data Processing
# ----------------------------------------------------------------------------
pandas>=2.0.0
numpy>=1.24.0,<2.3.0  # opencv-python과 호환성
scipy>=1.10.0

# ----------------------------------------------------------------------------
# Graph Processing
# ----------------------------------------------------------------------------
networkx>=3.1

# ----------------------------------------------------------------------------
# Geometric Operations (Plan B - Safety Metrics)
# ----------------------------------------------------------------------------
shapely>=2.0.0

# ----------------------------------------------------------------------------
# Computer Vision (SDD 데이터 처리)
# ----------------------------------------------------------------------------
opencv-python>=4.8.0

# ----------------------------------------------------------------------------
# Trajectory Data Processing (선택사항)
# ----------------------------------------------------------------------------
# trajdata>=1.0.0  # 필요시 주석 해제
# lanelet2>=1.1.0  # C++ 의존성 필요, 별도 설치 권장

# ----------------------------------------------------------------------------
# Visualization
# ----------------------------------------------------------------------------
matplotlib>=3.7.0
seaborn>=0.12.0

# ----------------------------------------------------------------------------
# Progress Bars
# ----------------------------------------------------------------------------
tqdm>=4.65.0

# ----------------------------------------------------------------------------
# Configuration Management
# ----------------------------------------------------------------------------
pyyaml>=6.0

# ----------------------------------------------------------------------------
# Logging and Monitoring
# ----------------------------------------------------------------------------
tensorboard>=2.13.0

# ----------------------------------------------------------------------------
# Machine Learning Utilities
# ----------------------------------------------------------------------------
scikit-learn>=1.3.0

# ----------------------------------------------------------------------------
# Jupyter Notebook Support
# ----------------------------------------------------------------------------
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=6.5.0

# ----------------------------------------------------------------------------
# Development Tools (선택사항)
# ----------------------------------------------------------------------------
# black>=23.0.0  # 코드 포맷팅
# flake8>=6.0.0  # 린터
# pytest>=7.0.0  # 테스트

# ============================================================================
# 설치 방법
# ============================================================================
# 1. 가상환경 생성:
#    python3 -m venv venv
#    source venv/bin/activate  # macOS/Linux
#    venv\Scripts\activate     # Windows
#
# 2. 패키지 설치:
#    pip install --upgrade pip
#    pip install -r requirements.txt
#
# 3. PyTorch Geometric 추가 설치 (필요시):
#    pip install torch-scatter torch-sparse torch-cluster torch-spline-conv \
#                -f https://data.pyg.org/whl/torch-2.0.0+cpu.html
#
# 4. Apple Silicon (M1/M2) 사용 시:
#    pip install torch torchvision torchaudio
#    # PyTorch Geometric은 위 명령으로 자동 설치됨
#
# 5. CUDA 사용 시:
#    pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
#    pip install torch-geometric torch-geometric-temporal
# ============================================================================
